Firefox schemas can be imported from schema/firefox-schemas
