+++ tests/schema/expected/manifest.json
@@ -1,0 +1,27 @@
+{
+  "id": "manifest",
+  "$ref": "#/types/WebExtensionManifest",
+  "permissions": [],
+  "definitions": {},
+  "refs": {},
+  "types": {
+    "Permission": {
+      "anyOf": [
+        {
+          "type": "string",
+          "enum": [
+            "alarms",
+            "clipboardRead",
+            "clipboardWrite",
+            "geolocation",
+            "idle",
+            "notifications",
+            "storage"
+          ]
+        },
+        { "$ref": "#/types/MatchPattern" },
+        { "$ref": "cookies#/definitions/Permission" }
+      ]
+    }
+  }
+}
